refactor(editor): Extract banners into features/shared (no-changelog)
