+++ packages/frontend/editor-ui/src/features/shared/banners/banners.types.ts
@@ -1,0 +1,9 @@
+import type { Component } from 'vue';
+import type { BannerName } from '@n8n/api-types';
+
+export type N8nBanners = {
+	[key in BannerName]: {
+		priority: number;
+		component: Component;
+	};
+};
