+++ packages/frontend/editor-ui/src/features/shared/banners/components/BannerStack.vue
@@ -1,13 +1,13 @@
 <script lang="ts">
-import NonProductionLicenseBanner from '@/components/banners/NonProductionLicenseBanner.vue';
-import TrialOverBanner from '@/components/banners/TrialOverBanner.vue';
-import TrialBanner from '@/components/banners/TrialBanner.vue';
-import V1Banner from '@/components/banners/V1Banner.vue';
-import EmailConfirmationBanner from '@/components/banners/EmailConfirmationBanner.vue';
-import DataTableStorageLimitWarningBanner from '@/components/banners/DataTableStorageLimitWarningBanner.vue';
-import DataTableStorageLimitErrorBanner from '@/components/banners/DataTableStorageLimitErrorBanner.vue';
+import NonProductionLicenseBanner from './banners/NonProductionLicenseBanner.vue';
+import TrialOverBanner from './banners/TrialOverBanner.vue';
+import TrialBanner from './banners/TrialBanner.vue';
+import V1Banner from './banners/V1Banner.vue';
+import EmailConfirmationBanner from './banners/EmailConfirmationBanner.vue';
+import DataTableStorageLimitWarningBanner from './banners/DataTableStorageLimitWarningBanner.vue';
+import DataTableStorageLimitErrorBanner from './banners/DataTableStorageLimitErrorBanner.vue';
 import type { Component } from 'vue';
-import type { N8nBanners } from '@/Interface';
+import type { N8nBanners } from '../banners.types';
 
 // All banners that can be shown in the app should be registered here.
 // This component renders the banner with the highest priority from the banner stack, located in the UI store.
