+++ packages/frontend/editor-ui/src/features/shared/banners/components/banners/BaseBanner.vue
