+++ packages/frontend/editor-ui/src/features/shared/banners/components/banners/DataTableStorageLimitErrorBanner.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import { useI18n } from '@n8n/i18n';
-import BaseBanner from '@/components/banners/BaseBanner.vue';
+import BaseBanner from './BaseBanner.vue';
 import { useDataTableStore } from '@/features/core/dataTable/dataTable.store';
 
 const dataTableStore = useDataTableStore();
@@ -8,10 +8,10 @@ const i18n = useI18n();
 </script>
 
 <template>
-	<BaseBanner name="DATA_TABLE_STORAGE_LIMIT_WARNING" :dismissible="true" theme="warning">
+	<BaseBanner name="DATA_TABLE_STORAGE_LIMIT_ERROR" :dismissible="true" theme="danger">
 		<template #mainContent>
 			<span>{{
-				i18n.baseText('dataTable.banner.storageLimitWarning.message', {
+				i18n.baseText('dataTable.banner.storageLimitError.message', {
 					interpolate: {
 						usage: `${dataTableStore.dataTableSize} / ${dataTableStore.maxSizeMB}MB`,
 					},
