+++ packages/frontend/editor-ui/src/features/shared/banners/components/banners/NonProductionLicenseBanner.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import BaseBanner from '@/components/banners/BaseBanner.vue';
+import BaseBanner from './BaseBanner.vue';
 import { i18n as locale } from '@n8n/i18n';
 </script>
