+++ packages/frontend/editor-ui/src/features/shared/banners/components/banners/TrialOverBanner.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import BaseBanner from '@/components/banners/BaseBanner.vue';
+import BaseBanner from './BaseBanner.vue';
 import { usePageRedirectionHelper } from '@/composables/usePageRedirectionHelper';
 import { i18n as locale } from '@n8n/i18n';
